Code now updating ping_target.lastsuccess, removing conditioncode column.
