+++ app/src/main/java/org/apache/roller/weblogger/business/PingResult.java
@@ -1,0 +1,59 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  The ASF licenses this file to You
+ * under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.  For additional information regarding
+ * copyright in this work, please see the NOTICE file in the top level
+ * directory of this distribution.
+ *
+ * Source file modified from the original ASF source; all changes made
+ * are also under Apache License.
+ */
+package org.apache.roller.weblogger.business;
+
+/**
+ * Conveys a ping result.
+ */
+public final class PingResult {
+
+    boolean error;
+    String message;
+
+    public PingResult(Boolean error, String message) {
+        this.error = error != null && error;
+        this.message = message != null ? message : "";
+    }
+
+    public boolean isError() {
+        return error;
+    }
+
+    public void setError(boolean error) {
+        this.error = error;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+
+    public String toString() {
+        return "PingResult{" + "error=" + error + ", message='" + message + "'" + "}";
+    }
+
+    // Inhibit construction
+    private PingResult() {
+    }
+}
